// src/components/parameters/dynamic-security-analysis/dynamic-security-analysis.type.ts
/**
 * Copyright (c) 2026, RTE (http://www.rte-france.com)
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */
export enum TabValues {
    SCENARIO = 'scenario',
    CONTINGENCY = 'contingency',
}